transactions db

+++ migrations/001_create_accounts_table.js
@@ -1,6 +1,6 @@
 exports.up = (pgm) => {
   pgm.createTable('accounts', {
     id: { type: 'varchar(255)', notNull: true, primaryKey: true },
-    balance: { type: 'bigint', notNull: true },
+    balance: { type: 'float', notNull: true },
   });
 };
