+++ migrations/002_create_transactions_table.js
@@ -3,7 +3,7 @@ exports.up = (pgm) => {
     id: { type: 'uuid', notNull: true, primaryKey: true },
     from: { type: 'varchar(255)', notNull: true },
     to: { type: 'varchar(255)', notNull: true },
-    amount: { type: 'bigint', notNull: true },
+    amount: { type: 'float', notNull: true },
     commited: {
       type: 'timestamp',
       notNull: true,
